+++ src/inputscope/conf.py
@@ -24,7 +24,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     26.03.2015
-@modified    13.04.2024
+@modified    15.04.2024
 ------------------------------------------------------------------------------
 """
 import ast
@@ -40,8 +40,8 @@ import sys
 
 """Program title, version number and version date."""
 Title = "InputScope"
-Version = "1.11.dev17"
-VersionDate = "13.04.2024"
+Version = "1.11.dev18"
+VersionDate = "15.04.2024"
 
 """TCP port of the web user interface."""
 WebHost = "localhost"