+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 - make keyboard heatmap image path configurable
 - retain heatmap replay close-button after completion
 - add program configuration path to database info box in statistics
+- add journaling file size to database file size in statistics info box
 - fix logging Numpad operator keys in Linux (issue #30)
 - fix SQLite-related warnings in Python 3.12+ (issue #30)
 - work around heatmap.js bug of not rendering keyboard heatmap replay if step > 1